+++ test/api-model.js
@@ -57,6 +57,9 @@ describe('Linz has a model api', function () {
                     },
                     sortBy: ['username', 'email'],
                 },
+                model: {
+                    title: 'username'
+                },
                 form: {
                     username: {
                         fieldset: 'Fieldset'